database_service/pyproject.toml
[project]
name = "database-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.15.1",
    "asyncpg>=0.30.0",
    "fastapi>=0.115.11",
    "greenlet>=3.1.1",
    "pydantic>=2.10.6",
    "python-dotenv>=1.0.1",
    "ruff>=0.11.0",
    "sqlalchemy>=2.0.39",
    "uvicorn>=0.34.0",
]
